+++ config/medium/Hunger/code/CGCode.js
@@ -1,0 +1,12 @@
+var inputs = readline().split(' ');
+const n = parseInt(inputs[0]);
+const k = parseInt(inputs[1]);
+var inputs = readline().split(' ');
+for (let i = 0; i < n; i++) {
+    const swt = parseInt(inputs[i]);
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('Feast duration');