+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "Thermal Flying - Episode 1".
 - configuration for "Guessing n Cheating".
 - configuration for "Dice galaxy".
+- configuration for "Hunger".
 
 ## [1.9.0] - 2024-11-30
 ### Added